+++ contracts/VM.sol
@@ -4,13 +4,13 @@ pragma solidity ^0.8.11;
 
 import "./CommandBuilder.sol";
 
-uint8 constant FLAG_CT_DELEGATECALL = 0x00;
-uint8 constant FLAG_CT_CALL = 0x01;
-uint8 constant FLAG_CT_STATICCALL = 0x02;
-uint8 constant FLAG_CT_VALUECALL = 0x03;
-uint8 constant FLAG_CT_MASK = 0x03;
-uint8 constant FLAG_EXTENDED_COMMAND = 0x80;
-uint8 constant FLAG_TUPLE_RETURN = 0x40;
+uint256 constant FLAG_CT_DELEGATECALL = 0x00;
+uint256 constant FLAG_CT_CALL = 0x01;
+uint256 constant FLAG_CT_STATICCALL = 0x02;
+uint256 constant FLAG_CT_VALUECALL = 0x03;
+uint256 constant FLAG_CT_MASK = 0x03;
+uint256 constant FLAG_EXTENDED_COMMAND = 0x80;
+uint256 constant FLAG_TUPLE_RETURN = 0x40;
 
 uint256 constant SHORT_COMMAND_FILL = 0x000000000000FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF;
 
@@ -34,7 +34,7 @@ abstract contract VM {
       internal returns (bytes[] memory)
     {
         bytes32 command;
-        uint8 flags;
+        uint256 flags;
         bytes32 indices;
 
         bool success;
@@ -43,7 +43,7 @@ abstract contract VM {
         uint256 commandsLength = commands.length;
         for (uint256 i; i < commandsLength; i=_uncheckedIncrement(i)) {
             command = commands[i];
-            flags = uint8(bytes1(command << 32));
+            flags = uint256(uint8(bytes1(command << 32)));
 
             if (flags & FLAG_EXTENDED_COMMAND != 0) {
                 indices = commands[i++];
